Move validate method into form code
