+++ controllers/apply/get-advice/docs/schema.test.js
@@ -4,14 +4,16 @@
 const fs = require('fs');
 const path = require('path');
 
-const schema = require('../schema');
+const formBuilder = require('../form');
+
+const form = formBuilder();
 
 const docContents = fs.readFileSync(
     path.resolve(__dirname, './schema.md'),
     'utf8'
 );
 
-test.each(Object.keys(schema.describe().children))(
+test.each(Object.keys(form.schema.describe().children))(
     '%s has documentation entry',
     fieldName => {
         const fieldNameRegexp = new RegExp(`### ${fieldName}`);
