+++ controllers/apply/get-advice/form.test.js
@@ -5,11 +5,7 @@ const omit = require('lodash/omit');
 const random = require('lodash/random');
 const faker = require('faker');
 
-const schema = require('./schema');
-
-function validate(data) {
-    return schema.validate(data, { abortEarly: false });
-}
+const formBuilder = require('./form');
 
 function mapMessages(validationResult) {
     return validationResult.error.details.map(detail => {
@@ -47,7 +43,8 @@ function mockResponse(overrides) {
 }
 
 test('minimal valid form', () => {
-    const result = validate(mockResponse());
+    const form = formBuilder();
+    const result = form.validate(mockResponse());
     expect(result.error).toBeNull();
     expect(result.value).toMatchSnapshot({
         projectIdea: expect.any(String),
@@ -56,7 +53,9 @@ test('minimal valid form', () => {
 });
 
 test('minimal invalid form', () => {
-    const result = validate({
+    const form = formBuilder();
+
+    const result = form.validate({
         projectCountry: 'invalid-country',
         projectLocation: null,
         projectCosts: 5000,
@@ -67,7 +66,9 @@ test('minimal invalid form', () => {
 });
 
 test('strip location and duration when applying for more than one country', () => {
-    const result = validate(
+    const form = formBuilder();
+
+    const result = form.validate(
         mockResponse({
             projectCountry: ['england', 'scotland'],
             projectLocation: 'this-should-be-stripped',
@@ -85,7 +86,9 @@ test.each([
     ['scotland', 3, 5],
     ['wales', 1, 5]
 ])('project duration in %p is %p–%p years', function(country, min, max) {
-    const result = validate(
+    const form = formBuilder();
+
+    const result = form.validate(
         mockResponse({
             projectCountry: country,
             projectDurationYears: random(min, max)
@@ -93,7 +96,7 @@ test.each([
     );
     expect(result.error).toBeNull();
 
-    const resultMin = validate({
+    const resultMin = form.validate({
         projectCountry: country,
         projectDurationYears: min - 1
     });
@@ -104,7 +107,7 @@ test.each([
         ])
     );
 
-    const resultMax = validate({
+    const resultMax = form.validate({
         projectCountry: country,
         projectDurationYears: max + 1
     });
@@ -119,7 +122,9 @@ test.each([
 test.each([['projectIdea', 50, 500], ['organisationBackground', 50, 500]])(
     '%p must be within word-count',
     function(fieldName, min, max) {
-        const resultMin = validate(
+        const form = formBuilder();
+
+        const resultMin = form.validate(
             mockResponse({
                 [fieldName]: faker.lorem.words(min - 1)
             })
@@ -131,7 +136,7 @@ test.each([['projectIdea', 50, 500], ['organisationBackground', 50, 500]])(
             ])
         );
 
-        const resultMax = validate(
+        const resultMax = form.validate(
             mockResponse({
                 [fieldName]: faker.lorem.words(max + 1)
             })
@@ -150,7 +155,9 @@ test.each([
     'organisationTradingName',
     'contactPhone'
 ])('optional %p field', function(fieldName) {
+    const form = formBuilder();
+
     const expected = omit(mockResponse(), fieldName);
-    const result = validate(expected);
+    const result = form.validate(expected);
     expect(result.error).toBeNull();
 });
